+++ src/libstore-tests/derivation/resolution.cc
@@ -2,6 +2,7 @@
 #include <nlohmann/json.hpp>
 
 #include "nix/store/derivations.hh"
+#include "nix/store/derivation/resolution.hh"
 #include "nix/store/derivation-options.hh"
 #include "nix/store/downstream-placeholder.hh"
 #include "nix/store/parsed-derivations.hh"
