+++ src/libstore/include/nix/store/derivations.hh
@@ -272,48 +272,6 @@ OutputsAndOptPaths outputsAndOptPaths(const Derivation<Inputs, Output> & drv, co
 bool hasDynamicDrvDep(const std::set<SingleDerivedPath> & inputs);
 
 /**
- * Determine whether this derivation should be resolved before building.
- *
- * Resolution is needed when:
- * - Input-addressed derivations are deferred (depend on CA derivations)
- * - Content-addressed derivations have input drvs and are either:
- *   - Floating (non-fixed), which must always be resolved
- *   - Fixed, which can optionally be resolved when ca-derivations is enabled
- * - Impure derivations always need resolution
- * - Any input derivations have outputs from dynamic derivations
- */
-bool shouldResolve(const Full & drv);
-
-/**
- * Return the underlying basic derivation but with these changes:
- *
- * 1. Input drvs are emptied, but the outputs of them that were used
- *    are added directly to input sources.
- *
- * 2. Input placeholders are replaced with realized input store
- *    paths.
- */
-std::optional<Basic> tryResolve(const Full & drv, Store & store, Store * evalStore = nullptr);
-
-/**
- * Like the above, but instead of querying the Nix database for
- * realisations, uses a given mapping from input derivation paths +
- * output names to actual output store paths.
- */
-std::optional<Basic> tryResolve(
-    const Full & drv,
-    Store & store,
-    fun<std::optional<StorePath>(ref<const SingleDerivedPath> drvPath, const std::string & outputName)>
-        queryResolutionChain);
-
-/**
- * Convert a `Basic` derivation to a `Full` derivation.
- * The resulting derivation has empty input drvs since a `Basic`
- * derivation is already resolved.
- */
-Full unresolve(const Basic & drv);
-
-/**
  * Check that the derivation is valid and does not present any
  * illegal states.
  *
@@ -361,6 +319,7 @@ void checkInvariants(const Derivation<Inputs, Output> & drv, Store & store, cons
  *
  * @param store The store to use for path computation
  */
+void fillInOutputPaths(Basic & drv, Store & store);
 void fillInOutputPaths(Full & drv, Store & store);
 
 /**
